[+] 客户端 服务器连调 调通版本

+++ Common/NetMsg.Game.Tank/Src/InputCmd.cs
@@ -13,7 +13,7 @@ namespace NetMsg.Game.Tank {
         }
 
         public override void Deserialize(Deserializer reader){
-            key = reader.GetByte();
+            key = reader.GetInt();
             val1 = reader.GetInt();
             val2 = reader.GetInt();
         }
